Added Animation.
git-svn-id: https://love.svn.sf.net/svnroot/love/trunk@413 3494dbca-881a-0410-bd34-8ecbaf855390

+++ src/Font.cpp
@@ -13,7 +13,7 @@ namespace love
 
 	float Font::getHeight() const
 	{
-		return size;
+		return (float)size;
 	}
 	
 	float Font::getWidth(const std::string & line) const
